Changed terraform setup for aws lambda and data-collect function to use aws lambda
