--- .dockerignore
@@ -1,4 +1,0 @@
-node_modules
-.git
-.DS_Store
-.env
