--- Dockerfile
@@ -1,29 +1,0 @@
-# Use the official Node.js 20 image as the base image
-FROM node:20
-
-# Set the working directory in the container
-WORKDIR /app
-
-# Copy package.json and package-lock.json to the working directory
-COPY package*.json ./
-
-# Install dependencies
-RUN npm install
-
-# Copy the rest of the application code
-COPY . .
-
-# Copy the entrypoint script
-COPY docker-entrypoint.sh /usr/local/bin/docker-entrypoint.sh
-
-# Make the entrypoint script executable
-RUN chmod +x /usr/local/bin/docker-entrypoint.sh
-
-# Expose the port the app runs on
-EXPOSE 8000
-
-# Define the entrypoint script
-ENTRYPOINT ["/usr/local/bin/docker-entrypoint.sh"]
-
-# Define the command to run the application
-CMD ["node", "index.js"]
