--- docker-entrypoint.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-set -e
-
-# Run any pre-start commands here (e.g., database migrations, environment setup)
-
-# Start the application
-exec "$@"
