--- src/terraform/variables.tf
@@ -1,16 +1,0 @@
-variable "aws_access_key_id" {
-  description = "AWS Access Key ID"
-  type        = string
-  sensitive   = true
-}
-
-variable "aws_secret_access_key" {
-  description = "AWS Secret Access Key"
-  type        = string
-  sensitive   = true
-}
-
-variable "s3_bucket_name" {
-  description = "Name of the S3 bucket"
-  type        = string
-}
